Moved syntax to its own file
